Search functionality and multi-language support

+++ src/main/controllers/Article.class.php
@@ -24,23 +24,39 @@ class Article extends AppHeader {
     private $_model, $_route, $_view, $_subApp;
 
     private $_errors = [];
+    private $_routeParams;
 
     public function __construct(?string $title, bool $isHomepage = false) {
         $this->loadConfig();
         $this->loadUser();
 
-        $this->_model = new Model($title, $isHomepage);
+        $this->checkForAjax();
         $this->_route = Registry::get('route');
 
-        if (!empty($this->_route->getParams()['action']))
+        if ($this->ajax)
+            $this->parseAjaxRequest();
+
+        $this->_routeParams = $this->_route->getParams();
+
+        if (!empty($this->_routeParams['action']) && $this->_routeParams['action'] == 'history' && !empty($this->_routeParams['id'])) {
+            $this->requiredPermission = 'view_article_history';
+            $this->_routeParams['action'] = 'article';
+            $this->_model = new Model($title, $isHomepage, $this->_routeParams['id']);
+        }
+        else
+            $this->_model = new Model($title, $isHomepage);
+
+        if (!empty($this->_routeParams['action']))
             $this->takeActionMeasures();
+        else
+            $this->_subApp = 'article';
 
         if (!empty($this->_subApp))
             $this->_model->setSubApp($this->_model->getIsEditing() && $this->canAccess('edit_articles') ? 'edit' : $this->_subApp);
     }
 
     private function takeActionMeasures() : void {
-        switch ($this->_route->getParams()['action']) {
+        switch ($this->_routeParams['action']) {
             case 'edit':
                 $this->requiredPermission = 'edit_articles';
                 $this->_subApp = 'edit';
@@ -52,8 +68,17 @@ class Article extends AppHeader {
                 break;
             case 'print':
                 $this->requiredPermission = 'print_articles';
-                $this->hasSidebar = false;
+                $this->hasSidebar = !self::$user->hasPermission('print_articles');
                 $this->_subApp = 'print';
+                break;
+            case 'history':
+                $this->requiredPermission = 'view_article_history';
+                $this->_subApp = 'history';
+                $this->_model->createHistoryTable();
+                break;
+            case 'assigntranslations':
+                $this->requiredPermission = 'assign_translations';
+                $this->_subApp = 'translations';
                 break;
             default:
                 $this->_subApp = 'unknown';
@@ -86,12 +111,19 @@ class Article extends AppHeader {
         if (!empty($_POST) && $this->_subApp == 'edit')
             $this->handleEditRequest();
 
+        if ($this->ajax && !empty($this->parsedAJAXRequest['getlist'])) {
+            $this->_model->setRequest('getlist');
+            return;
+        }
+
         $this->setTitle($this->_model->getTitle());
 
         if ($this->_subApp == 'delete') {
             $this->delete();
             return;
         }
+
+        $this->_model->setSubApp($this->_subApp);
 
         $data = $this->_model->prepareData() + ['errors' => $this->_errors];
         $this->_view = Registry::get('gui')->parseAndGet($this->_model->determineView(), $data, false);
@@ -102,6 +134,8 @@ class Article extends AppHeader {
     }
 
     public function getSidebarItems() : array {
+        $sidebarLanguages = $this->_model->getSidebarLanguages();
+
         return [
             'article' => [
                 'label' => 'current_article',
@@ -112,15 +146,25 @@ class Article extends AppHeader {
                         'href' => self::$config->read('url') . '/index.php/article/' . $this->_model->getSafeTitle(),
                         'visible' => true
                     ],
-                    'print' => [
-                        'label' => 'print_article',
-                        'href' => self::$config->read('url') . '/index.php/article/' . $this->_model->getSafeTitle() . '/print',
-                        'visible' => self::$user->hasPermission('print_articles')
-                    ],
                     'edit' => [
                         'label' => 'edit_article',
                         'href' => self::$config->read('url') . '/index.php/article/' . $this->_model->getSafeTitle() . '/edit',
                         'visible' => self::$user->hasPermission('edit_articles')
+                    ],
+                    'view_history' => [
+                        'label' => 'view_history',
+                        'href' => self::$config->read('url') . '/index.php/article/' . $this->_model->getSafeTitle() . '/history',
+                        'visible' => self::$user->hasPermission('view_article_history')
+                    ],
+                    'assign_translations' => [
+                        'label' => 'assign_translations',
+                        'href' => self::$config->read('url') . '/index.php/article/' . $this->_model->getSafeTitle() . '/assigntranslations',
+                        'visible' => self::$user->hasPermission('assign_translations')
+                    ],
+                    'print' => [
+                        'label' => 'print_article',
+                        'href' => self::$config->read('url') . '/index.php/article/' . $this->_model->getSafeTitle() . '/print',
+                        'visible' => self::$user->hasPermission('print_articles')
                     ],
                     'delete' => [
                         'label' => 'delete_article',
@@ -129,13 +173,22 @@ class Article extends AppHeader {
                         'request_confirm' => true
                     ]
                 ]
+            ],
+            'languages' => [
+                'label' => 'languages',
+                'visible' => !empty($sidebarLanguages),
+                'items' => $sidebarLanguages
             ]
         ];
     }
 
     public function getAdditionalHeaders() : ?string {
-        return $this->_subApp == 'print' ? '
+        return $this->_subApp == 'print' && self::$user->hasPermission('print_articles') ? '
             <link href="https://fonts.googleapis.com/css?family=EB+Garamond" rel="stylesheet">
             <link rel="stylesheet" type="text/css" href="' . self::$config->read('url') . '/index.php/stylesheet/print" />' : null;
+    }
+
+    public function printJSON() : void {
+        echo $this->_model->generateJSON();
     }
 }
